+++ src/commands/about.js
@@ -9,7 +9,7 @@ class LimeCommandAbout extends LimeCommand {
   // Constructor
   constructor(lime) {
     // Super from command class
-    super(lime, { name: 'about', description: 'shows information about the engine' });
+    super(lime, { name: 'about', description: 'show information about the engine' });
 
     // Operations
     this.operations.set(0, () => banner());