+++ src/structs/equation.js
@@ -17,6 +17,11 @@ class LimeEquation {
     return this.solution.length;
   }
 
+  // Get runtime
+  get runtime() {
+    return this.solution[this.steps - 1].timestamp - this.solution[0].timestamp;
+  }
+
   // Get result
   get result() {
     return this.solution[this.steps - 1].data;
@@ -26,6 +31,9 @@ class LimeEquation {
 
   // Record function
   record(step) {
+    if (this.lime.config.promptShowSteps) {
+      console.log(step.data);
+    }
     this.solution.push(step);
   }
 }